# Optimize jstring for the common case when unescaping is not needed