@bergmark 2020-02-21
Thanks a lot @Yuras!
LGTM, but frankly I never dug deep into the performance of aeson. Does it look to you @phadej @Lysxia?